Add the no items view in the list view so that it is shown and animated when there are no items.

+++ TodoList/View/ListView.swift
@@ -12,17 +12,23 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        listViewModel.updateItem(item: item)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                listViewModel.updateItem(item: item)
+                            }
                     }
+                    .onDelete (perform: listViewModel.deleteItem)
+                    .onMove (perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete (perform: listViewModel.deleteItem)
-            .onMove (perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List")
         .navigationBarItems(leading: EditButton(), trailing: NavigationLink("Add", destination: AddView()))
     }
